Create a separate class for location task data
