+++ ground/src/main/java/com/google/android/ground/ui/datacollection/tasks/location/CaptureLocationTaskViewModel.kt
@@ -21,14 +21,13 @@ import android.location.Location
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.asLiveData
 import com.google.android.ground.model.geometry.Coordinates
+import com.google.android.ground.model.geometry.Point
+import com.google.android.ground.model.submission.LocationTaskData
 import com.google.android.ground.system.FINE_LOCATION_UPDATES_REQUEST
 import com.google.android.ground.system.LocationManager
 import com.google.android.ground.system.PermissionsManager
 import com.google.android.ground.system.SettingsManager
 import com.google.android.ground.ui.datacollection.tasks.AbstractTaskViewModel
-import com.google.android.ground.ui.datacollection.tasks.point.LatLngConverter.processCoordinates
-import java.math.RoundingMode
-import java.text.DecimalFormat
 import javax.inject.Inject
 import kotlinx.coroutines.flow.distinctUntilChanged
 import kotlinx.coroutines.flow.map
@@ -44,7 +43,7 @@ constructor(
 
   val locationUpdates: LiveData<String> =
     locationManager.locationUpdates
-      .map { it.toTaskData().displayText() }
+      .map { it.toTaskData().getDetailsText() }
       .distinctUntilChanged()
       .asLiveData()
 
@@ -58,25 +57,11 @@ constructor(
     locationManager.disableLocationUpdates()
   }
 
-  data class TaskData(
-    val coordinates: Coordinates,
-    val altitude: Double?, // in metres
-    val accuracy: Float? // in metres
-  )
-
   companion object {
-    private fun Location.toTaskData(): TaskData {
+    private fun Location.toTaskData(): LocationTaskData {
       val altitude = if (hasAltitude()) altitude else null
       val accuracy = if (hasAccuracy()) accuracy else null
-      return TaskData(Coordinates(latitude, longitude), altitude, accuracy)
-    }
-
-    private fun TaskData.displayText(): String {
-      val df = DecimalFormat("#.##")
-      df.roundingMode = RoundingMode.DOWN
-      return "${processCoordinates(coordinates)}\n\n" +
-        "Altitude: ${df.format(altitude)}m\n" +
-        "Accuracy: ${df.format(accuracy)}m"
+      return LocationTaskData(Point(Coordinates(latitude, longitude)), altitude, accuracy)
     }
   }
 }
